+++ MoreShipUpgrades/Configuration/LategameConfiguration.cs
@@ -39,9 +39,10 @@ namespace MoreShipUpgrades.Configuration
     [DataContract]
     public class LategameConfiguration : SyncedConfig2<LategameConfiguration>
     {
+        public ITierEffectUpgradeConfiguration<int> SmarterLockpickUpgradeConfiguration { get; set; }
         public ITierEffectUpgradeConfiguration<int> EffectiveBandaidsConfiguration { get; set; }
         public ITierEffectUpgradeConfiguration<int> MedicalNanobotsConfiguration { get; set; }
-        public ITierEffectUpgradeConfiguration<int> ScrapKeeperConfiguration { get; set; }
+        public ScrapKeeperUpgradeConfiguration ScrapKeeperConfiguration { get; set; }
         public ITierEffectUpgradeConfiguration<int> ParticleInfuserConfiguration { get; set; }
         public IOneTimeUpgradeConfiguration SilverBulletsConfiguration { get; set; }
         public ITierCollectionUpgradeConfiguration FusionMatterConfiguration { get; set; }
@@ -95,6 +96,8 @@ namespace MoreShipUpgrades.Configuration
         public NightVisionUpgradeConfiguration NightVisionUpgradeConfiguration { get; set; }
         public SickBeatsUpgradeConfiguration SickBeatsUpgradeConfiguration {  get; set; }
         public BetterScannerUpgradeConfiguration BetterScannerUpgradeConfiguration {  get; set; }
+        public ITierEffectUpgradeConfiguration<int> BabyPacifierUpgradeConfiguration { get; set; }
+        public ITierEffectUpgradeConfiguration<int> ItemDuplicatorUpgradeConfiguration { get; set; }
         public ContractsConfiguration ContractsConfiguration { get; set; }
 
         #region Attributes
@@ -239,6 +242,13 @@ namespace MoreShipUpgrades.Configuration
 
             #region Upgrades
 
+            topSection = SmarterLockpick.UPGRADE_NAME;
+            SmarterLockpickUpgradeConfiguration = new TierPrimitiveUpgradeConfiguration<int>(cfg, topSection, LguConstants.SMARTER_LOCKPICK_ENABLED_DESCRIPTION, SmarterLockpick.PRICES_DEFAULT)
+            {
+                InitialEffect = cfg.BindSyncedEntry(topSection, LguConstants.SMARTER_LOCKPICK_INITIAL_INCREASE_KEY, LguConstants.SMARTER_LOCKPICK_INITIAL_INCREASE_DEFAULT, LguConstants.SMARTER_LOCKPICK_INITIAL_INCREASE_DESCRIPTION),
+                IncrementalEffect = cfg.BindSyncedEntry(topSection, LguConstants.SMARTER_LOCKPICK_INCREMENTAL_INCREASE_KEY, LguConstants.SMARTER_LOCKPICK_INCREMENTAL_INCREASE_DEFAULT, LguConstants.SMARTER_LOCKPICK_INCREMENTAL_INCREASE_DESCRIPTION),
+            };
+
             topSection = EffectiveBandaids.UPGRADE_NAME;
             EffectiveBandaidsConfiguration = new TierIndividualPrimitiveUpgradeConfiguration<int>(cfg, topSection, LguConstants.EFFECTIVE_BANDAIDS_ENABLED_DESCRIPTION, EffectiveBandaids.DEFAULT_PRICES)
             {
@@ -254,7 +264,7 @@ namespace MoreShipUpgrades.Configuration
             };
 
             topSection = ScrapKeeper.UPGRADE_NAME;
-            ScrapKeeperConfiguration = new TierPrimitiveUpgradeConfiguration<int>(cfg, topSection, LguConstants.SCRAP_KEEPER_ENABLED_DESCRIPTION, ScrapKeeper.PRICES_DEFAULT)
+            ScrapKeeperConfiguration = new ScrapKeeperUpgradeConfiguration(cfg, topSection, LguConstants.SCRAP_KEEPER_ENABLED_DESCRIPTION, ScrapKeeper.PRICES_DEFAULT)
             {
                 InitialEffect = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_KEY, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_DEFAULT, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_DESCRIPTION),
                 IncrementalEffect = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_KEY, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_DEFAULT, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_DESCRIPTION)
@@ -716,6 +726,20 @@ namespace MoreShipUpgrades.Configuration
 
             topSection = WalkieGPS.UPGRADE_NAME;
             WalkieGpsConfiguration = new OneTimeIndividualUpgradeConfiguration(cfg, topSection, LguConstants.WALKIE_GPS_ENABLED_DESCRIPTION, LguConstants.WALKIE_GPS_PRICE_DEFAULT);
+
+            topSection = BabyPacifier.UPGRADE_NAME;
+            BabyPacifierUpgradeConfiguration = new TierPrimitiveUpgradeConfiguration<int>(cfg, topSection, LguConstants.BABY_PACIFIER_ENABLED_DESCRIPTION, BabyPacifier.PRICES_DEFAULT)
+            {
+                InitialEffect = cfg.BindSyncedEntry(topSection, LguConstants.BABY_PACIFIER_INITIAL_PERCENTAGE_KEY, LguConstants.BABY_PACIFIER_INITIAL_PERCENTAGE_DEFAULT, LguConstants.BABY_PACIFIER_INITIAL_PERCENTAGE_DESCRIPTION),
+                IncrementalEffect = cfg.BindSyncedEntry(topSection, LguConstants.BABY_PACIFIER_INCREMENTAL_PERCENTAGE_KEY, LguConstants.BABY_PACIFIER_INCREMENTAL_PERCENTAGE_DEFAULT, LguConstants.BABY_PACIFIER_INCREMENTAL_PERCENTAGE_DESCRIPTION)
+            };
+
+            topSection = ItemDuplicator.UPGRADE_NAME;
+            ItemDuplicatorUpgradeConfiguration = new TierPrimitiveUpgradeConfiguration<int>(cfg, topSection, LguConstants.ITEM_DUPLICATOR_ENABLED_DESCRIPTION, ItemDuplicator.PRICES_DEFAULT)
+            {
+                InitialEffect = cfg.BindSyncedEntry(topSection, LguConstants.ITEM_DUPLICATOR_INITIAL_PERCENTAGE_KEY, LguConstants.ITEM_DUPLICATOR_INITIAL_PERCENTAGE_DEFAULT, LguConstants.ITEM_DUPLICATOR_INITIAL_PERCENTAGE_DESCRIPTION),
+                IncrementalEffect = cfg.BindSyncedEntry(topSection, LguConstants.ITEM_DUPLICATOR_INCREMENTAL_PERCENTAGE_KEY, LguConstants.ITEM_DUPLICATOR_INCREMENTAL_PERCENTAGE_DEFAULT, LguConstants.ITEM_DUPLICATOR_INCREMENTAL_PERCENTAGE_DESCRIPTION)
+            };
 
             #endregion