+++ MoreShipUpgrades/Patches/RoundComponents/RoundManagerPatcher.cs
@@ -6,6 +6,7 @@ using MoreShipUpgrades.UpgradeComponents.TierUpgrades.Ship;
 using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.UpgradeComponents.TierUpgrades;
 using MoreShipUpgrades.UpgradeComponents.TierUpgrades.Items;
+using System.Reflection.Emit;
 
 namespace MoreShipUpgrades.Patches.RoundComponents
 {
@@ -35,9 +36,7 @@ namespace MoreShipUpgrades.Patches.RoundComponents
             return codes;
         }
 
-        [HarmonyPatch(nameof(RoundManager.FinishGeneratingLevel))]
-        [HarmonyPrefix]
-        static void FinishGeneratingLevelPrefix()
+        public static void FinishGeneratingLevelPrefix()
         {
             if (!UpgradeBus.Instance.PluginConfiguration.LandingThrustersConfiguration.Enabled) return;
             if (!UpgradeBus.Instance.PluginConfiguration.LandingThrustersConfiguration.AffectLanding) return;
@@ -46,7 +45,20 @@ namespace MoreShipUpgrades.Patches.RoundComponents
 			StartOfRound.Instance.shipAnimator.speed *= LandingThrusters.GetLandingSpeedMultiplier();
         }
 
-        [HarmonyPatch(nameof(RoundManager.DespawnPropsAtEndOfRound))]
+        [HarmonyPatch(nameof(RoundManager.GenerateNewLevelClientRpc))]
+        [HarmonyTranspiler]
+        static IEnumerable<CodeInstruction> GenerateNewLevelClientRpcTranspiler(IEnumerable<CodeInstruction> instructions)
+        {
+            List<CodeInstruction> codes = new(instructions);
+            int index = 0;
+
+            FieldInfo spawnScrapAndEnemies = typeof(SelectableLevel).GetField(nameof(SelectableLevel.spawnEnemiesAndScrap));
+            Tools.FindField(ref index, ref codes, spawnScrapAndEnemies, skip: true);
+            codes.Insert(index - 2, new CodeInstruction(OpCodes.Call, typeof(RoundManagerPatcher).GetMethod(nameof(RoundManagerPatcher.FinishGeneratingLevelPrefix))));
+            return codes;
+        }
+
+		[HarmonyPatch(nameof(RoundManager.DespawnPropsAtEndOfRound))]
         [HarmonyTranspiler]
         static IEnumerable<CodeInstruction> DespawnPropsAtEndOfRoundTranspiler(IEnumerable<CodeInstruction> instructions)
         {
@@ -57,7 +69,10 @@ namespace MoreShipUpgrades.Patches.RoundComponents
             List<CodeInstruction> codes = new(instructions);
             int index = 0;
             Tools.FindField(ref index, ref codes, findField: IsScrap, addCode: CanKeepScrapBasedOnChance, andInstruction: true, notInstruction: true);
-            return codes;
+			codes.Insert(index, new CodeInstruction(opcode: OpCodes.Ldelem_Ref));
+			codes.Insert(index, new CodeInstruction(opcode: OpCodes.Ldloc, operand: 9));
+			codes.Insert(index, new CodeInstruction(opcode: OpCodes.Ldloc_0));
+			return codes;
         }
     }
 }